requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
pytest>=7.4.0

